+++ app/src/main/java/org/simple/clinic/login/pin/LoginPinScreenKey.kt
@@ -1,17 +1,15 @@
 package org.simple.clinic.login.pin
 
-import android.os.Parcelable
+import androidx.fragment.app.Fragment
 import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.android.parcel.Parcelize
-import org.simple.clinic.R
-import org.simple.clinic.router.screen.FullScreenKey
+import org.simple.clinic.navigation.v2.ScreenKey
 
 @Parcelize
-class LoginPinScreenKey : FullScreenKey, Parcelable {
+class LoginPinScreenKey : ScreenKey() {
 
   @IgnoredOnParcel
   override val analyticsName = "Login PIN Entry"
 
-  override fun layoutRes() = R.layout.screen_login_pin
-
+  override fun instantiateFragment() = LoginPinScreen()
 }